// src/payload-types.ts
export interface Post {
  id: string;
  title: string;
  heroImage?: (string | null) | Media;
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  relatedPosts?: (string | Post)[] | null;
  categories?: (string | Category)[] | null;
  meta?: {
    title?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (string | null) | Media;
    description?: string | null;
  };
  publishedAt?: string | null;
  authors?: (string | User)[] | null;
  populatedAuthors?:
    | {
        id?: string | null;
        name?: string | null;
      }[]
    | null;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
// ...
export interface Category {
  id: string;
  title: string;
  slug?: string | null;
  slugLock?: boolean | null;
  parent?: (string | null) | Category;
  breadcrumbs?:
    | {
        doc?: (string | null) | Category;
        url?: string | null;
        label?: string | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface ContentBlock {
  sectionGradient: 'none' | 'topToBottom' | 'bottomToTop' | 'middle' | 'radialCircleGreen';
  backgroundColor?:
    | (
        | '--background'
        | '--foreground'
        | '--foreground-secondary'
        | '--card'
        | '--card-foreground'
        | '--popover'
        | '--popover-foreground'
        | '--primary'
        | '--primary-foreground'
        | '--secondary'
        | '--secondary-foreground'
        | '--muted'
        | '--muted-foreground'
        | '--accent'
        | '--accent-foreground'
        | '--destructive'
        | '--destructive-foreground'
        | '--border'
        | '--input'
        | '--ring'
        | '--radius'
        | '--success'
        | '--warning'
        | '--error'
        | '--green'
        | '--accent-green'
      )
    | null;
  addSpotlight?: boolean | null;
  disableGutter?: boolean | null;
  columns?:
    | {
        size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
        maxWidth?: string | null;
        center?: boolean | null;
        verticalAlignment?: ('top' | 'center' | 'bottom') | null;
        topPadding?: ('none' | '4' | '8' | '12' | '16' | '20' | '24') | null;
        bottomPadding?: ('none' | '4' | '8' | '12' | '16' | '20' | '24') | null;
        richText?: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        blocks?:
          | (
              | FormBlock
              | MediaBlock
              | SectionNumberBlock
              | ContentCardSliderBlock
              | ContentToImageVerticalScrollBlock
              | ContentWithImageMiddleBlock
              | FAQBlock
              | CompareBlock
              | GalleryBlock
              | ProductArchiveBlock
              | OpenImageButtonBlock
              | CardBlock
              | ContentStructuralDetailsBlock
              | ConfigurationOptionsBlock
              | StructuralDiagramGalleryBlock
            )[]
          | null;
        enableLink?: boolean | null;
        link?: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?:
            | ({
                relationTo: 'pages';
                value: string | Page;
              } | null)
            | ({
                relationTo: 'posts';
                value: string | Post;
              } | null)
            | ({
                relationTo: 'products';
                value: string | Product;
              } | null);
          url?: string | null;
          label: string;
          /**
           * Choose how the link should be rendered.
           */
          appearance?: ('default' | 'outline') | null;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'content';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FormBlock".
 */
export interface FormBlock {
  form: string | Form;
  enableIntro?: boolean | null;
  formVariant?: ('variant1' | 'variant2') | null;
  contactInfos?:
    | {
        infoType: 'email' | 'phone' | 'address';
        richText: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
      }[]
    | null;
  introContent?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'formBlock';
}
// ...
export interface MediaBlock {
  media: string | Media;
  isParallax?: boolean | null;
  isExpandable?: boolean | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'mediaBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "SectionNumberBlock".
 */
export interface SectionNumberBlock {
  Index: number;
  id?: string | null;
  blockName?: string | null;
  blockType: 'sectionNumberBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentCardSliderBlock".
 */
export interface ContentCardSliderBlock {
  content: {
    contentBlocks: {
      contentBlock?: {
        root: {
          type: string;
          children: {
            type: string;
            version: number;
            [k: string]: unknown;
          }[];
          direction: ('ltr' | 'rtl') | null;
          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
          indent: number;
          version: number;
        };
        [k: string]: unknown;
      } | null;
      id?: string | null;
    }[];
    mediaBlocks: {
      media: string | Media;
      id?: string | null;
    }[];
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'contentCardSliderBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentToImageVerticalScrollBlock".
 */
export interface ContentToImageVerticalScrollBlock {
  content: {
    contentBlocks: {
      contentBlock?: {
        root: {
          type: string;
          children: {
            type: string;
            version: number;
            [k: string]: unknown;
          }[];
          direction: ('ltr' | 'rtl') | null;
          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
          indent: number;
          version: number;
        };
        [k: string]: unknown;
      } | null;
      id?: string | null;
    }[];
    mediaBlocks: {
      media: string | Media;
      id?: string | null;
    }[];
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'contentVerticalScrollBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentWithImageMiddleBlock".
 */
export interface ContentWithImageMiddleBlock {
  content: {
    contentBlocks: {
      contentBlock?: {
        root: {
          type: string;
          children: {
            type: string;
            version: number;
            [k: string]: unknown;
          }[];
          direction: ('ltr' | 'rtl') | null;
          format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
          indent: number;
          version: number;
        };
        [k: string]: unknown;
      } | null;
      id?: string | null;
    }[];
    mediaBlocks: {
      media: string | Media;
      id?: string | null;
    }[];
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'contentWithImageMiddleBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FAQBlock".
 */
export interface FAQBlock {
  faqItems: {
    heading: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    content: {
      root: {
        type: string;
        children: {
          type: string;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    id?: string | null;
  }[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'faqBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CompareBlock".
 */
export interface CompareBlock {
  image1: string | Media;
  image2: string | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'compareBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "GalleryBlock".
 */
export interface GalleryBlock {
  images?:
    | {
        image: string | Media;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'galleryBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ProductArchiveBlock".
 */
export interface ProductArchiveBlock {
  headerContent1?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  headerContent2?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  populateBy?: ('collection' | 'selection') | null;
  relationTo?: 'products' | null;
  projectType?: ('residential' | 'commercial') | null;
  showAllProjectsLink?: boolean | null;
  limit?: number | null;
  selectedDocs?:
    | {
        relationTo: 'products';
        value: string | Product;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'product-archive';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "OpenImageButtonBlock".
 */
export interface OpenImageButtonBlock {
  buttonLabel: string;
  image: string | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'openImageButtonBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CardBlock".
 */
export interface CardBlock {
  image?: (string | null) | Media;
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'cardBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContentStructuralDetailsBlock".
 */
export interface ContentStructuralDetailsBlock {
  enabled?: boolean | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'contentStructuralDetailsBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ConfigurationOptionsBlock".
 */
export interface ConfigurationOptionsBlock {
  id?: string | null;
  blockName?: string | null;
  blockType: 'configurationOptionsBlock';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "StructuralDiagramGalleryBlock".
 */
export interface StructuralDiagramGalleryBlock {
  images: {
    media?: (string | null) | Media;
    id?: string | null;
  }[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'structuralDiagramGalleryBlock';
}